move fake declare protostar version to conftest
