+++ protostar/self/__init__.py
@@ -1,5 +1,8 @@
 from .protostar_compatibility_with_project_checker import (
+    CompatibilityCheckResult,
     DeclaredProtostarVersionProviderProtocol,
+    ProtostarCompatibilityWithProjectChecker,
     ProtostarVersion,
     parse_protostar_version,
 )
+from .protostar_directory import VersionManager
