+++ protostar/self/protostar_compatibility_with_project_checker_test.py
@@ -1,5 +1,6 @@
 import pytest
 
+from .conftest import FakeDeclaredProtostarVersionProvider
 from .protostar_compatibility_with_project_checker import (
     CompatibilityCheckResult,
     DeclaredProtostarVersionProviderProtocol,
@@ -8,21 +9,13 @@ from .protostar_compatibility_with_project_checker import (
 )
 
 
-class FakeDeclaredProtostarVersionProvider(DeclaredProtostarVersionProviderProtocol):
-    def __init__(self, declared_protostar_version_str: str):
-        self._declared_protostar_version_str = declared_protostar_version_str
-
-    def get_declared_protostar_version(self):
-        return parse_protostar_version(self._declared_protostar_version_str)
-
-
 @pytest.fixture(name="declared_protostar_version_provider")
 def declared_protostar_version_provider_fixture(declared_protostar_version: str):
     return FakeDeclaredProtostarVersionProvider(declared_protostar_version)
 
 
 @pytest.mark.parametrize(
-    "protostar_version, declared_protostar_version, is_compatible",
+    "protostar_version, declared_protostar_version, check_result",
     (
         ("0.1.2", "0.1.2", CompatibilityCheckResult.COMPATIBLE),
         ("0.1.2", "0.1.1", CompatibilityCheckResult.COMPATIBLE),
@@ -36,7 +29,7 @@ def declared_protostar_version_provider_fixture(declared_protostar_version: str)
 def test_compatibility(
     declared_protostar_version_provider: DeclaredProtostarVersionProviderProtocol,
     protostar_version: str,
-    is_compatible: bool,
+    check_result: bool,
 ):
     compatibility_checker = ProtostarCompatibilityWithProjectChecker(
         parse_protostar_version(protostar_version),
@@ -45,4 +38,4 @@ def test_compatibility(
 
     result = compatibility_checker.check_compatibility()
 
-    assert result == is_compatible
+    assert result == check_result
